+++ provision-commands/install-mysql.sh
@@ -1,0 +1,9 @@
+# /*=============================
+# =            MYSQL            =
+# =============================*/
+sudo debconf-set-selections <<< 'mysql-server mysql-server/root_password password root'
+sudo debconf-set-selections <<< 'mysql-server mysql-server/root_password_again password root'
+sudo apt-get -y install mysql-server
+sudo mysqladmin -uroot -proot create scotchbox
+sudo apt-get -y install php7.2-mysql
+reboot_webserver_helper