+++ provision-commands/install-php.sh
@@ -1,0 +1,92 @@
+# /*===================================
+# =            INSTALL PHP            =
+# ===================================*/
+
+# Install PHP
+sudo add-apt-repository -y ppa:ondrej/php # Super Latest Version (currently 7.2)
+sudo apt-get update
+sudo apt-get install -y php7.2
+sudo apt-get -y install libapache2-mod-php
+
+# Add index.php to readable file types
+MAKE_PHP_PRIORITY='<IfModule mod_dir.c>
+    DirectoryIndex index.php index.html index.cgi index.pl index.xhtml index.htm
+</IfModule>'
+echo "$MAKE_PHP_PRIORITY" | sudo tee /etc/apache2/mods-enabled/dir.conf
+
+sudo service apache2 restart
+
+
+# /*===================================
+# =            PHP MODULES            =
+# ===================================*/
+
+# Base Stuff
+sudo apt-get -y install php7.2-common
+sudo apt-get -y install php7.2-dev
+
+# Common Useful Stuff (some of these are probably already installed)
+sudo apt-get -y install php7.2-bcmath
+sudo apt-get -y install php7.2-bz2
+sudo apt-get -y install php7.2-cgi
+sudo apt-get -y install php7.2-cli
+sudo apt-get -y install php7.2-fpm
+sudo apt-get -y install php7.2-gd
+sudo apt-get -y install php7.2-imap
+sudo apt-get -y install php7.2-intl
+sudo apt-get -y install php7.2-json
+sudo apt-get -y install php7.2-mbstring
+sudo apt-get -y install php7.2-odbc
+sudo apt-get -y install php-pear
+sudo apt-get -y install php7.2-pspell
+sudo apt-get -y install php7.2-tidy
+sudo apt-get -y install php7.2-xmlrpc
+sudo apt-get -y install php7.2-zip
+
+# Enchant
+sudo apt-get -y install libenchant-dev
+sudo apt-get -y install php7.2-enchant
+
+# LDAP
+sudo apt-get -y install ldap-utils
+sudo apt-get -y install php7.2-ldap
+
+# CURL
+sudo apt-get -y install curl
+sudo apt-get -y install php7.2-curl
+
+# IMAGE MAGIC
+sudo apt-get -y install imagemagick
+sudo apt-get -y install php7.2-imagick
+
+
+
+
+
+# /*===========================================
+# =            CUSTOM PHP SETTINGS            =
+# ===========================================*/
+PHP_USER_INI_PATH=/etc/php/7.2/apache2/conf.d/user.ini
+
+
+echo 'display_startup_errors = On' | sudo tee -a $PHP_USER_INI_PATH
+echo 'display_errors = On' | sudo tee -a $PHP_USER_INI_PATH
+echo 'error_reporting = E_ALL' | sudo tee -a $PHP_USER_INI_PATH
+echo 'short_open_tag = On' | sudo tee -a $PHP_USER_INI_PATH
+sudo service apache2 restart
+
+# Disable PHP Zend OPcache
+echo 'opache.enable = 0' | sudo tee -a $PHP_USER_INI_PATH
+
+# Absolutely Force Zend OPcache off...
+sudo sed -i s,\;opcache.enable=0,opcache.enable=0,g /etc/php/7.2/apache2/php.ini
+sudo service apache2 restart
+
+
+# /*================================
+# =            PHP UNIT            =
+# ================================*/
+sudo wget https://phar.phpunit.de/phpunit-6.1.phar
+sudo chmod +x phpunit-6.1.phar
+sudo mv phpunit-6.1.phar /usr/local/bin/phpunit
+reboot_webserver_helper