+++ provision-commands/install.sh
@@ -1,0 +1,1 @@
+/provision-commands/provision-lamp.sh