+++ provision-commands/intall-wpcli.sh
@@ -1,0 +1,6 @@
+# /*==============================
+# =            WP-CLI            =
+# ==============================*/
+curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+sudo chmod +x wp-cli.phar
+sudo mv wp-cli.phar /usr/local/bin/wp