+++ provision-commands/update-machine.sh
@@ -1,0 +1,15 @@
+# Update packages indexes
+sudo apt-get update
+
+# The following is "sudo apt-get -y upgrade" without any prompts
+sudo DEBIAN_FRONTEND=noninteractive apt-get -y -o Dpkg::Options::="--force-confdef" -o Dpkg::Options::="--force-confold" upgrade
+
+sudo apt-get install -y build-essential
+sudo apt-get install -y tcl
+sudo apt-get install -y software-properties-common
+sudo apt-get install -y python-software-properties
+sudo apt-get -y install vim
+sudo apt-get -y install git
+
+# Weird Vagrant issue fix
+sudo apt-get install -y ifupdown